breakdown workflow cli more
